Add fake insulin log dao

+++ app/src/main/java/com/axel_stein/glucose_tracker/data/model/InsulinLog.kt
@@ -10,7 +10,7 @@ data class InsulinLog(
     @ColumnInfo(name = "insulin_id")
     val insulinId: Long,
 
-    val units: Int,
+    val units: Float,
 
     val measured: Int,
